NIX[HOME]: Switch terminal emulator to Alacritty.
Sakura's search function is borderline useless, while Alacritty's is
quite serviceable. Since it looks like Alacritty can be themed
identically, it's an easy switch. The only real complication is that it
doesn't support VTE integration like Sakura does, but that is easily
worked around with a small edit to $PROMT_COMMAND.